Tidying up yesterday's code. JSON objects are used in the callback functions for getStory to print outlines, some (very) minor styling applied to mainTest.php
Tidying up yesterday's code: Some array issues with the getStory block
in main.php, moved all SQL (that’s in use) to queries.php, some sanity
checking and tidying up of calls to echo.
JSON objects are used in the callback functions for getStory to print
outlines. The components can be moused over to show a bootstrap tooltip
which has more information in.
Some (very) minor styling applied to mainTest.php: Params and php
echoes are now stored in a modal to avoid cluttering the page.
I also added the arc_desc and arc_es attributes to the updateES
function of the Person class and parse the arc_desc to show an updating
mood in the character modal.