Created a Story class for saving and processing stories, added a block for evaluating the stories to mainTest.php
Created a Story class for saving and processing stories: Includes new
queries to get the latest story ID and update the story rating - as
well as methods in Story and Main to handle this.
Added a block for evaluating the stories to mainTest.php
Moved the layout of mainTest.php around a bit
Added a random event/location seed to the selectors
Added a break_early variable to stop adding to the element/location
lists if respect_death is set and a character has died
Added long descriptions for consequences to action and event classes -
also display these in hint text when mousing over the action/event
consequence.
Changed some variable names to be more consistent