More tweaks to the user evaluation page, some quick hacks to make getting evaluated story data easier
More tweaks to the user evaluation page: Changed the explanatory text a
bit, also plan to start hosting a copy on a publicly accessible host to
send to people for feedback tomorrow or Thursday.
Some quick hacks to make getting evaluated story data easier: Added
some dataSet toggles and a button to spit the evaluated stories and
their metadata out to the console in a format that is easy to use as
variables in an arraylist. (This is a bit bananas I know, but it was a
quick hack to get the data out and the user evaluation proof of concept
working.)
Also call the mainSetup function at the end of a story request to
refresh the values from the KB and stopped drawing charts: I plan on
moving the charts to their own page and re-doing the queries that make
them as they’re not quite relevant to the data I want to gather.