Split the charts into their own module, finished up the user evaluation site
Split the charts into their own module: removed (some) of the code from
the maintTest.php module and into stats.html which will contain the
data when I figure out what exactly that should be - hopefully in the
next day or so when I get the first of my user feedback. This also
includes a new section in Main.php - with some repeated code (echoing
variables) that could probably be tidied up.
Finished up the user evaluation site: Re-wrote the text to explain the
project and what’s being asked clearly, added the final (first?) set of
stories (6 of each dataset).
Updated the insert statement generator.