Culled the user evaluation list from 24 to 12, tweaked the charts and their labels, aded Levenshtein distance calculator for assessing the similarity of stories
Culled the user evaluation list from 24 to 12: Removed 3 from each
dataset to make the user evaluations shorter.
Tweaked the charts and their labels: Now have slightly better colours
and am using positively evaluated and negatively evaluated not good and
bad.
Added Levenshtein distance calculator for assessing the similarity of
stories: The distance is calculated between a random set of 50
sequences chosen from the KB and the new story sequence. This is then
averaged and compared. Each distance can be added (event, location,
action) and averaged for the whole stories lev distance from the KB and
therefore potential fitness. This is only comparing to positively
evaluated stories atm, as the negative kb data is essentially nil.
A really quick page to check the story Levenshtein distance for each
story by id was also added…